DopEFW

Documentation for DopEFW.

DopEFW.param_color

DopEFW.param_marker

DopEFW.Formula

DopEFW.bootstrap

DopEFW.df_bootstrap

DopEFW.get_sheet_names

DopEFW.indices_names

DopEFW.read_excel_sheets

DopEFW.simulation

DopEFW.split_string

DopEFW.param_color — Constant

Dict containing the colors for each parameter.

DopEFW.param_marker — Constant

Dict containing the markers for each parameter.

DopEFW.Formula — Type

struct Formula

A structure to represent a mathematical formula or function along with its parameters and associated paper.

Fields

func::Function: A function that defines the mathematical operation or formula.

parameters::Vector{String}: A vector of strings representing the names of the parameters used in the formula.

paper::String: A string indicating the source or reference (e.g., paper title) associated with the formula.

Example

formula_instance = Formula(my_func, ["x", "y"], "Title of the Paper")

DopEFW.bootstrap — Method

bootstrap(data::Vector{T}, n_samples::Int, stat_func::Function, seed::Int) where T

Performs bootstrap resampling on a given dataset to compute statistics.

Arguments

data::Vector{T}: A vector of data points from which bootstrap samples will be drawn.

n_samples::Int: The number of bootstrap samples to generate.

stat_func::Function: A function that takes a vector of samples and returns a statistic (e.g., mean, median).

seed::Int: An integer seed for the random number generator to ensure reproducibility.

Returns

A vector of bootstrap statistics computed from the generated samples.

DopEFW.df_bootstrap — Method

df_bootstrap(simulations::DataFrame, formula::Formula; n_samples = 1000, stat_func = mean, seed = 1234)

Performs bootstrap resampling on a given formula from data from the simulation dataframe to compute statistics.

Arguments

simulations::DataFrame: A dataframe containing the simulations.

formula::Formula: The Formula object that will be bootstrapped.

n_samples = 1000: The number of bootstrap samples to generate.

stat_func = mean: A function that takes a vector of samples and returns a statistic (e.g., mean, median).

seed = 1234: The random seed

Returns

A vector of bootstrap statistics computed from the generated samples.

DopEFW.get_sheet_names — Method

get_sheet_names(file_path::String) -> Vector{String}

Retrieves the names of all sheets in an Excel file.

Arguments

file_path::String: The path to the Excel file from which to read the sheet names.

Returns

A vector of strings containing the names of the sheets in the specified Excel file.

Example

sheet_names = get_sheet_names("example.xlsx")
println(sheet_names)  # Output: ["Sheet1", "Sheet2", "Data", ...]

DopEFW.indices_names — Method

indices_names(formula::Formula) -> Vector{String}

Generates names for the Sobol sensitivity indices based on the parameters defined in a given Formula.

Arguments

formula::Formula: An instance of the Formula struct, which contains the function, parameters, and associated paper.

Returns

A vector of strings representing the names of the Sobol sensitivity indices. This includes combinations of the parameters and total order indices.

Example

formula_instance = Formula(x -> x[1] + x[2], ["x", "y", "z"], "Title of the Paper")

index_names = indices_names(formula_instance)
println(index_names)  # Output: ["x", "y", "z", "x_y", "x_T", "y_T", "z_T"]

DopEFW.read_excel_sheets — Method

read_excel_sheets(file_path::String) -> DataFrame

Reads all sheets from an Excel file and combines them into a single DataFrame.

Arguments

file_path::String: The path to the Excel file from which to read the sheets.

Returns

A DataFrame containing the combined data from all sheets in the specified Excel file. Each sheet's data is transformed and includes an additional column indicating the sheet name.

DopEFW.simulation — Method

simulation(formula::Formula, dataset::String; ci::Int = 10, path::String = "DopEFW\data\pro\data.csv", samples::Int = 10^6) -> DataFrame

Performs a sensitivity analysis simulation using the specified formula and dataset, generating Sobol sensitivity indices.

Arguments

formula::Formula: An instance of the Formula struct that contains the function to evaluate and its parameters.

dataset::String: The name of the dataset (paper) to filter the data for the simulation.

ci::Int: The confidence interval percentile to use (default is 10).

path::String: The file path to the CSV file containing the datasets (default is "DopEFW\data\pro\data.csv").

samples::Int: The number of samples to use for the Quasi-Monte Carlo simulation (default is 1,000,000).

Returns

A DataFrame containing the results of the sensitivity analysis simulation, for each week of the dataset.

Example

# Define a simple function and create a Formula instance

formula_instance = Formula(x -> x[1] + x[2], ["x1", "x2"], "Title of the Paper")

results = simulation(formula_instance, "DatasetName")
println(results)

DopEFW.split_string — Method

split_string(s::String) -> Tuple{String, String}

Splits a string into two parts at the first occurrence of a digit.

Arguments

s::String: The input string to be split.

Returns

A tuple containing two strings:

The first element is the substring before the first digit.

The second element is the substring from the first digit to the end of the string. If no digit is found, the second element will be an empty string.

Example

result = split_string("Hello123World")
println(result)  # Output: ("Hello", "123World")